// examples/uss-omaha_batch-4/uss-omaha/src/types/content.ts
export interface Phases {
  heading: string
  phaseList: Phase[]
}

export interface Phase {
  number: number
  title: string
  description: string
  status: string
  cost: string
  percentComplete: number
}